Split and Functional
Script is split into two seperate files, one for parsing and one for
generation since parsing takes forever.
